MNISTDemo split
